refactor(doctor): use %q in format error message and table-driven tests
Use %q instead of %s in printDiagnosticsTo error message for
consistency with resolveFormat and defensiveness against unquoted
values in error output.

Convert statusLabel and statusEmoji individual test functions to
table-driven style per TC-006 SHOULD convention.

Move doctor renderer helper tests (resultLabel, statusEmoji,
countStatusSummary) from cli_test.go to doctor_test.go to co-locate
them with the other doctor format tests introduced in #744.

Addresses review feedback from #744.
